clean up typing indicator
